+++ CHANGELOG.md
@@ -3,6 +3,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## Added
+ - Added a new option to enable a good news (no breach) message, when an entry is manually checked ('check current password')
 
 ## [0.5.0] - 2020-04-26
 ### Added